Changed some funtions name.

+++ kernel/includes/fs.h
@@ -34,7 +34,7 @@ void load_root_entries();
 DirEntry_t find_file(char* name, char* extension);
 
 void load_file(char* buffer, DirEntry_t file);
-uint16_t get_next_cluster_number(uint16_t current_cluster);
+int get_next_cluster_number(int current_cluster);
 
 void* file_read(char* name, char* ext);
